Improved build script

+++ build.sh
@@ -1,4 +1,16 @@
 #!/bin/sh
-docker build -t go-docker .
-docker rm -f go-contacts
-docker run -d --name go-contacts -p 8080:8080 go-docker
+
+if [[ "$(docker images -q golang:1.16-alpine)" == "" ]]; then
+	docker build -t golang:1.16-alpine .
+fi
+
+if [[ "$(docker ps -q -f name=go-contacts)" != "" ]]; then
+	docker rm -f go-contacts
+fi
+
+docker run -d --name go-contacts -p 8080:8080 golang:1.16-alpine
+
+echo
+echo "Successful project build"
+echo "Open it on http://localhost:8080"
+echo
